requirements.txt
requests
minio
psycopg2-binary
apache-airflow-providers-jdbc
pyspark